requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
click==8.1.8
distro==1.9.0
fastapi==0.115.6
fireworks-ai==0.15.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
idna==3.10
jiter==0.8.2
loguru==0.7.3
openai==1.59.6
pillow==11.1.0
pydantic==2.10.5
pydantic-core==2.27.2
python-dotenv==1.0.1
python-slugify==8.0.4
sniffio==1.3.1
starlette==0.41.3
text-unidecode==1.3
tqdm==4.67.1
typing-extensions==4.12.2
uvicorn==0.34.0
wsproto==1.2.0
